Moves API to emoji-api.com
The previous API wasn't actually public and the owner changed it a few
times breaking this. This site's API is still in beta, but is intented
for programatic access. Others need to fill in an API key for this to
  work.

The responses from this API can be quite large, many hundreds of items
returned. This can slow down the Spoon.